chore(appContext): mv to common, compiles
