+++ package.json
@@ -8,7 +8,8 @@
     "bunyan": "^1.8.14",
     "font-awesome": "^4.7.0",
     "i18next": "^19.7.0",
-    "lit-html": "^1.2.1"
+    "lit-html": "^1.2.1",
+    "node-fetch": "^2.6.0"
   },
   "files": [
     "README.md",
