+++ src/@littleware/little-elements/common/appContext/i18n.ts
@@ -1,11 +1,12 @@
 import * as i18n from "../../../../../i18next/dist/esm/i18next.js";
+import AppContext from './appContext.js';
+
 
 /**
  * An AppContext-aware loader for 18next:
  * - https://i18next.com
  * - https://www.npmjs.com/package/i18next
  */
-
  class I18nHelper {
      /** 
       * See https://stackoverflow.com/questions/25606730/get-current-locale-of-chrome,
