+++ src/@littleware/little-elements/common/appContext/index.html
@@ -1,0 +1,28 @@
+{% extends "lib/styleGuide/shell/basicShell.html.njk" %}
+
+{% block head %}
+<title>App Context Canvas</title>
+
+<link rel="stylesheet" type="text/css" href="/modules/jasmine-core/lib/jasmine-core/jasmine.css" />
+<script src="/modules/jasmine-core/lib/jasmine-core/jasmine.js"></script>
+<script src="/modules/jasmine-core/lib/jasmine-core/jasmine-html.js"></script>
+<script src="../test/littleJasmineBoot.js"></script>
+{% endblock %}
+
+{% block content %}
+<section id="README">
+  {% markdown %}
+     {% include "lib/authMgr/README.md" %}
+  {% endmarkdown %}
+</section>
+
+<section>
+
+<p> Here is the lw-app-context custom element </p>
+<lw-app-context></lw-app-context>
+
+
+</section>
+<script src="./spec/main.js" type="module"></script>
+{% endblock %}
+
