+++ src/@littleware/little-elements/common/appContext/spec/appContextSpec.ts
@@ -1,0 +1,21 @@
+import { sleep } from "../../mutexHelper.js";
+import "../appContext.js";
+
+interface Tools {
+    i18n: any;
+    log: any;
+}
+
+describe( "the application context", () => {
+    let tools:Tools = null;
+
+    beforeAll((done) => {
+
+    });
+
+    it( "Can dance", () => {
+        expect(true).toBe(true);
+    });
+
+
+});
