+++ src/@littleware/little-elements/common/appContext/spec/main.ts
@@ -1,0 +1,4 @@
+import {startTest} from "../../../lib/test/util.js";
+import "./appContextSpec.js";
+
+startTest();
